directory restructuring cleanup scss vars remove unused files add routes alias
