+++ craco.config.js
@@ -3,6 +3,7 @@ module.exports = {
   webpack: {
     alias: {
       '@components': path.resolve(__dirname, 'src/components'),
+      '@routes': path.resolve(__dirname, 'src/routes'),
       '@utils': path.resolve(__dirname, 'src/utils'),
     },
   },
