+++ src/components/CollectionLog/LogEntryList.tsx
