+++ src/index.tsx
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom';
 
 import './index.scss';
 
-import Footer from './components/Footer/Footer';
-import Header from './components/Header/Header';
+import Footer from '@components/footer/Footer';
+import Header from '@components/header/Header';
 import Router from './router';
 
 ReactDOM.render(
