+++ src/scss/_variables.scss
@@ -15,7 +15,6 @@ $account-type-shadow: drop-shadow(1px 1px 0 black)
 $text-shadow: 2px 2px 0px black;
 
 $log-border: 4px solid $bg-darker;
-$log-border-image: url('../public/img/log_border.png') 10 repeat;
 $log-shadow: inset 0px 2px 0px 0px rgba(0,0,0,0.2);
 
 @font-face {
