+++ tsconfig.json
@@ -22,6 +22,7 @@
     "baseUrl": ".",
     "paths": {
       "@components/*": ["./src/components/*"],
+      "@routes/*": ["./src/routes/*"],
       "@utils/*": ["./src/utils/*"]
     }
   },
